image-processing-service/pyproject.toml
[tool.poetry]
name = "image-processing-service"
version = "0.1.0"
description = ""
authors = ["maver01 <matteo.sabadin4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.4.0"
kafka-python = "^2.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
